fix: user relationships
on trying to create mock users and wallets related to each one in the same seeding, an error was thrown because the relationship between the user and wallet model hasn't been specified in user model. I also noticed that the other relationships - like with tasks, notifications... - haven't been as well, so i've specified them too.