refactor: remove balance column from wallets table
there's no need at all for the column 'balance' to exist: it's possible to get the amount of money the user has from the sum of transactions made that are associated with their wallet. But also, since there's no other attribute in wallets despite the user fk and timestamps, there's also no need for a wallets table: the transactions are now associated directly to the user.