fix: syntax error to get transactionCategory id
Find is used to search for model instances by primary keys; actually the intention was to make a query based on a column value that is not a primary key. The right method for it is Model::where, and the 'first()' is needed because 'where()' returns a collection, and we just want the first element of this collection.