feat: caching on wallet data
since part of the derived data from wallet show view is not frequently changed and the query to retrieve it can potentially become slow  - nextBillDueDate and nextTaskDueDate -, it's now cached.
The 'balance' is not cached because can frequently change and full_name not too because it does not involve a query and it's got simply but an accessor which concatenates first_name and full_name, thus the difference would be negligible and sometimes even worse if caching is applied to full_name - due to the queries to store the cached value.