fix: correctly get the next pending task/bill due date
The queries before were fetching the tasks/bills which due_date was the latest one; now they fetch the data considering these conditions: the due_date has to be in the future, the status has to be 'pending' and from the result set get on these conditions, get the first on ascending order - that is: the nearest future pending bill due_date. The views were changed to comply with this fixing, and now consider that no data as return from the query is a possibility - 'none'.