fix: format date on 'none' error in wallet
if 'none' is stored in the cache record related to the nextPendingBillDueDate, then in the wallet.show view the '->format()' can be tried to be applied on 'none' if the query in walletController returns null. In order to avoid that error, now the query can return null and not 'none'.