fix/refactor: change comparison in handleCache listeners
instead of sending the event to the handler method to handled there based on the 'state' - if it was created, updated... - of the entity sent in the event, the event type is now used to determine where to send the event. This was avoiding the BillUpdateTest to work, because the BillUpdate event was being sent to 'handleUpdatedBill' because the entity's 'wasRecently'property was true and the wasChanged() returning false - the bill was being created on the test and updated, but indirectly: a request was being made to the bills.update route and then the bill was being updated there, but the route does not returns the updated bill and the '$bill' continues the same way as before; it only has been used as an argument to the route method. On refreshing the bill the 'wasCreated' continued returning false, so this refactor was made to fix a bug on testing update-related tests on entities such as bills, transactions and tasks.