feat: benchmark tests for wallet routes
benchmark class with tests for the wallet routes (in this case, there's at the moment only one route: show). The structure used for the benchmark tests is the following: a benchmark class with methods that execute a code that sends a request to an endpoint. It's important to add up that it's assured that the database changes after the request are not persisted at the database at the end of the method execution.