feat: add observers for Bill and Task
replacing the events and listeners for observers: instead of having events related to model actions (in this case create, update, delete) being dispatched manually in the controllers and listeners to listen to them, it's easier and appropriate to have observers that watch for these model actions and execute a function directly (before the dispatched would get into the appropriate listener in its 'handler', and in the 'handler' the proper other method from the listener would be executed with the event as an argument, such as 'handleCreatedBill') based on the model action triggered. Events and listeners are more appropriate for actions that are unrelated to the models.