fix: assertDatabase on tests failing due to difference between timestamps in the database x in model instance
when creating an entity instance using factories in tests and having to compare it with the data stored in the database, always an error would occur: the created_at and updated_at are automatically defined when the data comes to the database by default - no matter if we define them in the model instance. So there's no need and to put the timestamp fields in the factories.