fix: assert on due_date failing when comparing entity instance against database record
in tests involving comparing model instances with due_date field against the value stored in the database, the test would always fail due to the casting to 'datetime': after processing the data in the request and creating a model instance + record based on it, 'due_date' is casted to be a Carbon instance with datetime format - considering the time. Besides, there's no need at all for the database hold in due_date columns dates with time: just the date in 'Y-m-d' format is enough.