fix: transactions amount not according to type
it was possible to store a transaction of type 'expense' that has as amount field as a positive number, and one of type 'income' as a negative number. Also it was possible to add a '-' at the left of the number in the frontend input and this way, following the previous logic, an amount of type 'expense' would be a positive number (10 = --10). Now it's enforced that the values sent to the server will be considered always as absolute (positive form) and the logic of storing the amount properly based on the type properly works (it just now considers the type, since the amount is assured absolute).