fix: amount expected in db not equals to requestData in TransactionStoreTest
the requestData amount is not always the same as the one the database stored: there's now a trigger that makes the amount be stored either as a positive number or negative depending on the type of the transaction informed. So it's necessary to consider this behavior in the test to assert correctly the data that is expected after the post request using the requestData.