refactor: improve charts logic using more library related code
laravel charts provide ways to make the code cleaner on dealing with charts, one of them is that the chart can be shown in the view and mentioned via javascript with an exclusive syntax, avoiding the need to provide too much js code for showing and updating it. ps: one of the resources available there was related to ajax calls, but i haven't implemented it 'cause it does not convey a way to dynamically import the labels calculated in the chart, but only its data.