fix: bill factory adjusted to trigger logic related to paid status
mock bills are now not created invalid (that is, no way a bill which status is not 'paid' gets a value for paid_at field and a paid bill gets no value in paid_at). Also the paid_at cannot be greater than the current timestamp (which is going to be enforced by another trigger).