feat: restrict paid_at update triggers
the paid_at field of a record cannot be updated unless the status to or is already paid and if the date informed is in the future. The triggers logics have not yet been applied at the application level because 'paid_at' is not fillable at the moment (it will be futurely, since it's planned to allow the user to define and change paid_at in the app forms).