feat: add trigger after update transaction category type
i assured that a transaction only can be related to a transaction_category of the same type, but i haven't considered before the situation where a transaction_category type changes - which is not intended, but can happen. If that happens, now there's a trigger setting the transaction_type as NULL in all transactions related to that type in transactions table.