refactor: move all authorization and validation logic from controllers to Request classes
the authorization logic for enabling users to only tinker with their own data and also a the only restant validation on controllers that has no built-in rule in laravel (if the transaction_category sent on the request is not only a reference of a row from transaction_categories, but also if the row 'transaction_type' is equivalent to the transaction 'type' that we are updating|storing) have been moved to Request Form classes. Some of these classes didn't existed before and have been created to hold just authorization logic - since it would be confusing to have authorization logic in 'delete' controller method but not on 'create' or 'update'.