fix: redirect back to 404 after deleting in show view
if the url where the delete request was made from is a show endpoint, then the redirect isn't back: it navigates the user to the index endpoint. The reason why i just didn't left only redirect to the index endpoint is that, in the case of deleting something from an specific page in index pagination, then the previous location of the user in the pagination would be lost.