refactor: chart use and update purely in html/js
not anymore with the syntax provided by Laravel Charts, and considering data being returned from the new structure of ChartControllers -> ChartDataServices. Also the size of the chart has been increased in height to make it fill up in dashboard view all the place available aside from latest-transactiosn and wallet components.