chore: replace mailgun with postmark as mailing service
Mailgun in the free tier just enables emails to be sent to up to five different recipients, and also was giving me trouble on configuration: i couldn't register an email domain and the error message about it was quite generic and didn't shows any instruction or even why the problem was happening. The only advantage is the 5k free emails credit monthly available, but i found Postmark much better: configuration is pretty simple and almost everything is abstracted from you, the UI is more user-friendly and emails can be sent to anyone. The free tier only enables 100 emails per month, but thinking about the scope of this project and long-term it's suitable and better than mailgun.