refactor: pass query options explicitly to views by index methods
The 'index' method in transaction/bill/task controller didn't sent before any query options to index views: the 'filter/sortBy' values being used in QueryOption classes and in the index views was coming directly from the request, and the values, especially of 'filterBy', were not consistent: if just one checkbox 'filterBy' was selected, the value passed to the request would be primitive; if more, then it would be an array: that difference was considered both in QueryOption and views. Now the filterBy is always an array, providing more consistency, less code *in general* without damaging readability - in general because the controller size increased, but at least the approach more straightforwardly informs the developer which filters and sorts are applied to the view. The ideia of refactoring this way came from filtering, but also on sorting the same logic has been applied to enforce code consistency.