Revert/Refactor "Merge pull request #65 from augusto-dmh/refactor/query-options-explicitly-passed-to-views-by-index-methods-and-always-as-arrays"
'filterByType' was lacking the '[]' aside from it in transactions index view, then the is_array was necessary there, but not at all on the other filters. So i added the lacking '[]' and removed the unnecessary 'is_array' checks; by that always 'filterBy' will be arrays. Now i only need to check for the availability of 'filterBy' for the scenario of a request without filters - both in index views and QueryOption classes. So i think the advantages from the last refactor do not all worth compared to these: more code in the controllers and QueryOptions to avoid a nullable checking in the views? I don't think so: even the filters applied are already straightforward in the controller - not specifying if the QueryOption is either for filtering or sorting, but still i see this is as enough. Nevertheless, the selects default options have been manteined.