fix: query to get the transaction_category most used in transactions returning null unwantedly
the transactions were grouped based on transaction_category_id, but the column is nullable, so it was possible that the group returned was from the transactions with transaction_category_id as null: that resulted in crashing the application on dashboard route. Now the the fields in this column with null as value are not anymore considered in the query.