feat: disable triggers + simulate 'insert' ones in TestUserSeeder
There are insert triggers that, when executed on test data, create inconsistencies (e.g: when a paid bill is inserted, a transaction is created based on it and the 'created_at' field receives the current timestamp, but considering that the bill that originate the transaction can have a random 'paid_at'/'due_date', that is ilogic). So that's why the triggers are disabled and the behavior of insert ones are simulated instead, avoiding these inconsistencies.