feat: make created_at fillable in Transactions to fill in factories
the transactions made are showed in the dashboard chart  in a range which max is ten years, and the column used to determine when the transaction was made is 'created_at'. So these changes aim to make the test data generated for the application more consistent and also visible in the dashboard chart.