fix: correctly avoid triggers in mysql before factories run in TestUserSeeder
the usability of triggers in mysql can't rely on variables. It's possible to take an approach where a table is created to have a row that controlls whether the trigger action expected by the trigger would be executed or not (the table row would be checked before anything in all triggers), but i'd rather another: drop the triggers and add them again after the factories have run in TestUserSeeder.