fix: apply searching only if searchTerm is sent
Currently the searching will only be applied if a searchTerm was sent: filtering and searching are separated and fetching results based on both is not intended - it'll be added in future updates. Fix made to be possible to fetch filtered data especially for bills, which is going to be used in 'BillsOverdue' email as a link to them.