fix: corret query logic in controllers to fully integrate searching+filtering+sorting
the query before was allowing, due to badly used boolean expressions, to fetch data inconsistent with the filtering applied (e.g: filtering for bills with status 'completed' and searching for bills with in title or description the searchTerm 'Labore non' and getting a bill that is not 'completed' as part of the result set). This problem has been fixed.