chore: remove sidebar dynamic change to avoid responsivity problems
the sidebar was capable of dynamically change, when a button is pressed, the margin of the main element, thus the distribution of elements inside: all the view design is damaged depending on the device used. So this dynamic behaviour has been removed to avoid the need to add tailwindcss classes dynamically based on the width of the 'main' element, or defining in 'lg' devices the same distribution of elements as in 'sm'.