style: corrected centering of 'show' views and success flash-message
changes have been made to ensure that the flash message of success (for now only used in 'bills/show' ) and the main element in 'show' views will be centered properly. To achieve this, the flash message has been placed in layout view instead of one specific, and the "page wrapper" and the wrapper around "header" and "main" elements in in 'layout' view have been changed.