fix: add status input and correct due date naming to follow convention
it was verified in one of the tests that one of the fillable properties from the Bill model didn't have two correspondent inputs in create view. One needed to be added and other was not named accordingly to the convention i've now declared: "all create/update views will have inputs which correspond to all the fillable attributes from a model, except the foreign key and automatically set ones".