fix: correct bill due_date caching logic on bill deleted
before when a bill whose due_date was stored in the cache was deleted, the key stored (user_[bill->user_id]_next_bill_due) would be assigned to a null value rather than deleted. Now the logic has changed: after the deletion, if there's a bill that matches the necessary conditions, its due_date is cached; if no, then the key stored gets deleted.

This problem was identified when writting the bill-due-date-caching related tests - not yet commited, but will be in BillStoreTest.