chore: add comment on 'show' method from BillController on redirect logic
if the bill deletion request is sent when the previous url points to bill 'show' endpoint, then the user is redirect to 'index' endpoint. If it'd be a redirect back, in this case the user would back to a 'not found' page - since there's no bill with that id anymore (show endpoint: bills/bill/id).