feat: add index and savePreferences methods to UserAvailableNotificationController
the form data needs to be sent to only one endpoint, so there needs to be a method that enables both bulkStore and bulkDestroy of notifications enabled by the user: savePreferences. The 'store' and 'destroy' were used because i thought previously on having eventListeners attached to the checkboxes with these methods involved, but this approach would trigger one request per click, which is worse to performance than just one request-response cycle that deals with all the notifications storing/deletion.