feat: add mainContainerRowDirection prop to layout view
variable to control the stylying of the "content" of the layout view (Page Main Container and Page Content), making it centered in the middle of the view and with header and main aside to each other. Thus, there's a new way to customize the position of the "content" in views.