fix: modify visibility of Notification classes properties to public
they wouldn't be accessible in the tests when accesing them to test the content ($notification->bills). These fixes of the commit are related to the notification-content tests actually, but made when the ideia was to test both the notification content and the channel in sole tests; i decided later to separate these assertions into different ones.