fix: add 'paid_at' field to Table\row|columns component
i thought i had a problem with the a method from BillChartDataService to get the last paid bills but i hadn't: i've looked at the table in Bill\index view and see there were paid bills in the last seven days there weren't showed in the chart, and confirmed it wasn't a problem related to the table: the database data was according to it. Well, actually i had looked at 'due date' instead of 'paid at', and noticed that no 'paid at' was even mentioned in the table, so i added it up there (by modifying both row and column components) to avoid get confused anymore.